// frontend/src/app/models/Contenedor.ts
export interface Contenedor{
    matricula?: number;
    idCalle?: string;
    numCalle?: string;
    idActividad?: string;
    numOrden?: string;
    idCapacidad?: string;
    idProveedor?: string;
    situacion?: string;
    estadoretorn?: string;
    fechinatala?: any;
    denomActividad?: string;
    solicitante?: string;
    DNI_NIF?: string;
    fechaUltMov?: Date;
    fechaBaja?: Date;
    observacionesRetorn?: string;
    expedienteAlta?: string;
    cargaLateral?: boolean;
    caseta?: boolean;
    soterrado?: boolean;
}