// frontend/src/app/models/Reposicion.ts
export interface Reposicion{
    matricula?: number;
    fecha?: Date;
    expediente?: string;
    entregadoPor?: string;
    idCalle?: number;
    numCalle?: string;
    idActividad?: string;
    numOrden?: string;
    causa?: string;
    DNIS?: string;
    receptor?: string;
    DNIR?: string;
    asignatario?: string;
    DNIA?: string;
}